miniscript: add an optional Pk to PkH nodes
This allows to access the actual public key behind the pubkey_hash. This
can be used to derive public keys behind pubkey_hash in Miniscripts with
extended public keys by modifying the translation function.

However, the preimage is not available when parsing Miniscript from
Script so it is inherently Option'al.

Signed-off-by: Antoine Poinsot <darosior@protonmail.com>